requirements.txt
plotly==5.20.0
dash==2.16.1
numpy==1.26.4
git+https://github.com/openai/CLIP.git
dash-bootstrap-components
pandas
dash-ag-grid
wordcloud
dash-image-gallery
dash-core-components
umap-learn
scikit-learn
wget
dash-holoniq-wordcloud
cleanlab-studio
python-dotenv